Move all tabManager tracking code to a separate component
