+++ shared/js/background/startup.js
@@ -71,16 +71,6 @@ export async function onStartup () {
         showContextMenuAction()
     }
 
-    const savedTabs = await browser.tabs.query({ status: 'complete' })
-    for (let i = 0; i < savedTabs.length; i++) {
-        const tab = savedTabs[i]
-
-        if (tab.url) {
-            // On reinstall we wish to create the tab again
-            await tabManager.restoreOrCreate(tab)
-        }
-    }
-
     if (resolveReadyPromise) {
         resolveReadyPromise()
         resolveReadyPromise = null
